Commonize SVGCanvasTest with non-jvm targets
